lead-qualifier/requirements.txt
fastapi
uvicorn[standard]
openai
pandas
scikit-learn
lightgbm
joblib
python-dotenv
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openpyxl
scipy
gradio
sqlalchemy
google-genai
psycopg2-binary
alembic
